9 gateway microservices

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interaction/client/shopping/cart/ShoppingCartClientFallbackFactory.java
@@ -22,7 +22,7 @@ public class ShoppingCartClientFallbackFactory implements FallbackFactory<Shoppi
             }
 
             @Override
-            public ShoppingCartDto addProductToShoppingCart(String username, Map<UUID, Integer> products) {
+            public ShoppingCartDto addProductToShoppingCart(String username, Map<UUID, Long> products) {
                 throw new RuntimeException("Shopping-cart service unavailable", cause);
             }
